Fold getApiProxyUrl into request-context.ts
Both helpers answer routing-decision questions for the provider layer
(proxy URL + context-injection gate), and the separate file wasn't
buying isolation — providers-server.ts already imports from
request-context.ts. One file, same logic.
